Add basic ToDo list
Still adding functionality. Cribbed most of this from my earlier react-todo project
so it made things pretty quick. Going to look into how to enable a mongoDB to save
the Todo items between sessions. COULD use LocalStorage but I want to figure out
MongoDB and this is a good excuse