Add a NodeJS-Express-Mongoose backend
Big commit! Meant to chunk this into smaller ones but got carried away trying this stuff out.
This adds a second NodeJS server via ExpressJS. server.js is the entry point and I've got it set up
to accept Todo items. I'm writing out the delegate now to keep things clean.

Currently it hooks up into a locally-hosted MongoDB. It's a bit of a pain to start two Node servers
AND a MongoDB, so I think I'm going to Dockerize this. I'll keep that as a separate branch though, I think
it should be pretty easy to do, being said.